Created dedicated sub-package to isolation functions. Moved spherical_isolation into it and associated testing
