+++ halotools/mock_observables/__init__.py
@@ -20,6 +20,7 @@ from .tpcf_multipole import *
 from .error_estimation_tools import jackknife_covariance_matrix
 from .pairwise_velocity_stats import *
 from .isolation_criteria import *
+from .isolation_functions import *
 from .void_stats import *
 from .catalog_analysis_helpers import *
 from .pair_counters import (npairs_3d, npairs_projected, npairs_xy_z,
