+++ halotools/mock_observables/isolation_functions/__init__.py
@@ -1,0 +1,6 @@
+""" Sub-package of `~halotools.mock_observables` providing the 
+isolation criteria functionality. 
+"""
+
+from .spherical_isolation import spherical_isolation
+
